services/ekchat-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
asyncpg==0.29.0
greenlet==3.1.1
python-jose[cryptography]==3.3.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
openai==1.3.7
azure-storage-blob==12.24.0
redis==5.0.1
hiredis==2.2.3
python-dateutil==2.8.2
openpyxl==3.1.5
matplotlib==3.9.2
pypdf==5.2.0
python-docx==1.1.2
reportlab==4.2.5
